requirements.txt
python-slugify
crossref_commons
python-magic-bin
